requirements.txt
flask
flask-cors
openai-whisper
ffmpeg-python
gunicorn
